services/inference-worker/services/requirements.txt
# =============================================================================
# Common dependencies shared across services
# Core packages pinned, ecosystem packages use compatible ranges
# Last updated: 2026-01-26
# =============================================================================

# -----------------------------------------------------------------------------
# Core Utilities (pinned)
# -----------------------------------------------------------------------------
requests>=2.32.0,<3.0.0
python-dotenv>=1.0.0
tqdm>=4.66.0
tabulate>=0.9.0
jq>=1.8.0
msgpack>=1.0.0
anyio>=4.0.0
pyyaml>=6.0.0
aiohttp>=3.9.0

# -----------------------------------------------------------------------------
# Data Processing (pinned for stability)
# -----------------------------------------------------------------------------
pandas>=2.2.0,<3.0.0
numpy>=1.26.0,<2.0.0
lxml>=5.0.0
pyarrow>=18.0.0

# -----------------------------------------------------------------------------
# AWS SDK (compatible range)
# -----------------------------------------------------------------------------
boto3>=1.35.0
botocore>=1.35.0

# -----------------------------------------------------------------------------
# LLM and RAG Core Packages - LangChain Ecosystem
# Using compatible ranges to avoid conflicts
# -----------------------------------------------------------------------------
openai>=1.50.0
tiktoken>=0.7.0

# LangChain Core (0.3.x branch)
langchain>=0.3.0,<0.4.0
langchain-core>=0.3.0,<0.4.0
langchain-text-splitters>=0.3.0,<0.4.0
langchain-openai>=0.3.0,<0.4.0
langchain-anthropic>=0.3.0,<0.4.0
langchain-community>=0.3.0,<0.4.0
langchain-experimental>=0.3.0,<0.4.0

# LangGraph
langgraph>=0.2.0
langgraph-checkpoint>=2.0.0
langgraph-sdk>=0.1.0

# LangSmith
langsmith>=0.2.0

# Memory management for agents
langmem>=0.0.20

# -----------------------------------------------------------------------------
# Observability and Tracing (Phoenix/OpenInference)
# -----------------------------------------------------------------------------
arize-phoenix-otel>=0.12.0
openinference-instrumentation-anthropic>=0.1.15
openinference-instrumentation-langchain>=0.1.50
openinference-instrumentation-openai>=0.1.20

# -----------------------------------------------------------------------------
# Message Queue and Database
# -----------------------------------------------------------------------------
pika>=1.3.0,<2.0.0
motor>=3.3.0,<4.0.0
pymongo>=4.6.0,<5.0.0
sqlalchemy>=2.0.0,<3.0.0
redis>=5.0.0,<6.0.0

# -----------------------------------------------------------------------------
# Web Framework (for inference service)
# -----------------------------------------------------------------------------
fastapi>=0.115.0
uvicorn>=0.30.0

# -----------------------------------------------------------------------------
# Communication Integrations
# -----------------------------------------------------------------------------
slack-sdk>=3.30.0
gitpython>=3.1.40

# -----------------------------------------------------------------------------
# Pydantic (data validation)
# -----------------------------------------------------------------------------
pydantic>=2.9.0,<3.0.0
pydantic-settings>=2.5.0,<3.0.0
